move helper to utils
